chore: move the vnode types under the main directory
