+++ packages/@lwc/engine-core/src/framework/component.ts
@@ -5,9 +5,10 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
 import { assert, isFalse, isFunction, isUndefined } from '@lwc/shared';
-import { invokeComponentRenderMethod, isInvokingRender, invokeEventListener } from './invoker';
+
+import { VNodes } from './vnode';
 import { VM, scheduleRehydration } from './vm';
-import { VNodes } from '../3rdparty/snabbdom/types';
+import { invokeComponentRenderMethod, isInvokingRender, invokeEventListener } from './invoker';
 import { ReactiveObserver } from '../libs/mutation-tracker';
 import { LightningElementConstructor } from './base-lightning-element';
 import { Template, isUpdatingTemplate, getVMBeingRendered } from './template';
